% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_AC.R
\name{diff_AC}
\alias{diff_AC}
\title{Identifying differential tRNA-index}
\usage{
diff_AC(
  AC_count,
  condition,
  replicates,
  LFC = 0,
  Padj = 0.05,
  color = c("orange", "purple", "grey")
)
}
\arguments{
\item{AC_count}{The path to the Anticodon_counts_raw.txt file output by mim-tRNA-seq
e.g AC_count <- "Anticodon_counts_raw.txt"}

\item{condition}{a list,Select the order of the experimental and control groups
in the Anticodon_counts_raw.txt file and list them.
e.g condition <- list(N1 = c(1,2), N2 = c(3,4))
Selecting the experimental group as 1 and 2, and the control group as 3 and 4 from the Anticodon_counts_raw.txt file}

\item{replicates}{a logical, Does the tRNA-seq data have biological replicates? If yes, set replicates = T,
if no, set replicates = F.}

\item{LFC}{The log-fold change (LFC) for differential tRNA between two groups, default is 0}

\item{Padj}{Threshold for significant differential tRNAs, default is 0.05}

\item{color}{Volcano plot color settings for different tRNA-indexes}
}
\description{
Identifying differential tRNA-index
}
\examples{
## AC_count <- "../Anticodon_counts_raw.txt"
## condition <- list(N1 = c(3,4),N2 = c(5,6))
## replicates = T
## tRNA_res <- diff_AC(AC_count,condition,T)

}
