% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPA_res.R
\name{EPA_res}
\alias{EPA_res}
\title{Extract codon information at E, P, and A sites for each read from ribo-seq data.}
\usage{
EPA_res(cdna_path, reads_psite_list, geneinfo, tx_scale = "the_longest")
}
\arguments{
\item{cdna_path}{Corresponding species cDNA.fa file}

\item{reads_psite_list}{List file generated by the psite_info function of riboWaltz}

\item{geneinfo}{The output of geneinfo, a Genome annotation matrix}

\item{tx_scale}{If transcripts were not filtered in the previous steps, here the transcript with the
longest CDS will be used to represent the gene.If it is already one gene to one transcript, tx_scale = "no".
default is "the_longest"}
}
\description{
Extract codon information at E, P, and A sites for each read from ribo-seq data.
}
\examples{

## EPA_result<-EPA_res(cdna_path = "../Mus_musculus.GRCm39.cdna.all.fa",reads_psite_list = reads_psite_list,geneinfo,tx_scale ="no")

}
