% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneinfo.R
\name{extractGeneInfo}
\alias{extractGeneInfo}
\title{Genome annotation matrix}
\usage{
extractGeneInfo(gtf.path)
}
\arguments{
\item{gtf.path}{Corresponding species gtf annotation file}
}
\description{
Genome annotation matrix
}
\examples{

## geneinfo <- extractGeneInfo("../Mus_musculus.GRCm39.110.chr.gtf")

}
