% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_peak.R
\name{TE_peak}
\alias{TE_peak}
\title{To calculate the translation efficiency of differentially expressed peaks (diff-peak) genes}
\usage{
TE_peak(
  RNA_count,
  CDS_count,
  gene,
  peak_pic,
  condition,
  replicates,
  gene_id = T,
  min_RNA = 10,
  min_CDS = 10,
  color = c("#BCE3A0", "#60D2D4")
)
}
\arguments{
\item{RNA_count}{Transcriptome result matrix, with the first column as gene ID/transcript ID,
followed by columns for gene counts in different samples.}

\item{CDS_count}{ribo-seq results matrix, with the first column being gene ID/transcript ID,
followed by several columns representing the RPF of different samples}

\item{gene}{The output of geneinfo, a Genome annotation matrix}

\item{peak_pic}{The output of tRNA_diff_peak}

\item{condition}{Two strings, The names of the experimental group and the control group.
(e.g. condition <- c("Heart","Liver"))}

\item{replicates}{If the number of biological replicates is the same for both the experimental and control groups,
input only one number. If they are different, input the number of biological replicates for both the
experimental and control groups separately (e.g replicates = c(2,3)).}

\item{gene_id}{If the first column of RNA_count and CDS_count is gene ID, then gene_id = T.
If it is transcript ID, then gene_id = F.default is gene_id = T}

\item{min_RNA}{The minimum threshold for the average gene expression level in the transcriptomic matrix.}

\item{min_CDS}{The minimum threshold for the average RPF level in the ribo-seq results matrix.}

\item{color}{The color of the bar chart}
}
\description{
To calculate the translation efficiency of differentially expressed peaks (diff-peak) genes
}
\examples{

## RNA<-read.table("../CDS/RNA.txt",sep = "\t",header = T)
## CDS<-read.table("../CDS/CDS.txt",sep = "\t",header = T)
## TE_pic<- TE_peak(RNA_count = RNA,CDS_count = CDS,gene = geneinfo,peak_pic = peak_pic ,condition = condition ,replicates = 2 ,gene_id = F)

}
