% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_diff_peak.R
\name{tRNA_diff_peak}
\alias{tRNA_diff_peak}
\title{Combine the results of EPA_res and diff_AC to identify the differential peaks caused by tRNA.}
\usage{
tRNA_diff_peak(
  EPA_path,
  condition,
  replicates,
  tRNA,
  geneinfo,
  min_diff = 3,
  diff_tRNA_num = 8,
  padj = 0.05,
  LFC = 0,
  min_count = 10,
  Non_translation_Elongation = 5,
  peak_method = "mean",
  only_both = F,
  bar_color = "#60D2D4",
  line_color = "#FE4A01"
)
}
\arguments{
\item{EPA_path}{a list, List the results of EPA_res in the order of the experimental group and the control group.
e.g if replicates = F  EPA_path <- list(S1 = list(EPA_result[[1]]), S2 = list(EPA_result[[2]]))
e.g if replicates = T  EPA_path <- list(S1 = list(EPA_result[[1]],EPA_result[[2]]),S2 = list(EPA_result[[3]],EPA_result[[4]]))}

\item{condition}{Two strings, The names of the experimental group and the control group.
(e.g. condition <- c("Heart","Liver"))}

\item{replicates}{a logical, Does the Ribo-seq data have biological replicates? If yes, set replicates = T,
if no, set replicates = F.}

\item{tRNA}{The output of diff_AC.}

\item{geneinfo}{The output of geneinfo, a Genome annotation matrix}

\item{min_diff}{Set the minimum peak ratio difference for the same gene at the same locus between two group.default is 3}

\item{diff_tRNA_num}{Set the significant difference in the number of tRNAs. If diff_tRNA_num = "all",
then input all significant different tRNAs. If the value of diff_tRNA_num is greater than
the number of significant different tRNAs, input all significant different tRNAs.
If the value of diff_tRNA_num is less than the number of significant different tRNAs,
input the diff_tRNA_num most significant tRNAs. default is 8}

\item{padj}{Threshold for significant differential tRNAs, default is 0.05}

\item{LFC}{The log-fold change (LFC) for differential tRNA between two groups, default is 0}

\item{min_count}{The minimum number of ribosome-protected fragments (RPFs) per gene, default is 10}

\item{Non_translation_Elongation}{Exclude peak judgments associated with translation initiation and termination sites.
Exclude peak results from regions downstream of start codons and upstream of stop codons by n codons. default is 5}

\item{peak_method}{Calculate peak values using two methods: mean, which computes the ratio of reads at each site to
the average reads per site across the gene; and median, which computes the ratio of reads at each site to the
median reads per site across the gene. default is mean}

\item{only_both}{Whether differential peak sites have reads in both groups, default is F}

\item{bar_color}{The colors of the bar chart, default is '#1F78B4'}

\item{line_color}{The colors of differential peak sites, default is '#FE4A01'}
}
\description{
Combine the results of EPA_res and diff_AC to identify the differential peaks caused by tRNA.
}
\examples{
## peak_pic <- tRNA_diff_peak(EPA_path = EPA_path,tRNA = tRNA_res,condition = condition,replicates = T,
##            geneinfo = geneinfo,min_diff = 8,diff_tRNA_num = 5)

}
